+++ scipy/special/_logiv.pxd
@@ -1,0 +1,1 @@
+cdef double log_iv_asym(double v, double z) nogil
