+++ scipy/special/setup.py
@@ -109,6 +109,11 @@ def configuration(parent_package='',top_path=None):
                          sources=['_ellip_harm_2.c', 'sf_error.c',],
                          **cfg)
 
+    # logiv
+    config.add_extension('_logiv',
+                         sources=['_logiv.c'],
+                         define_macros=define_macros)
+
     # Cython API
     config.add_data_files('cython_special.pxd')
 
@@ -138,11 +143,6 @@ def configuration(parent_package='',top_path=None):
     # combinatorics
     config.add_extension('_comb',
                          sources=['_comb.c'])
-
-    # logiv
-    config.add_extension('_logiv',
-                         sources=['_logiv.c'],
-                         define_macros=define_macros)
 
     # testing for _round.h
     config.add_extension('_test_round',
